refactor(swiftui): extract shared FicheroWebView, de-dup WKWebView wrappers (#1699)
